code/pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = "project code goes here."
authors = ["frankcholula <lu.phrank@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
scipy = "^1.14.1"
matplotlib = "^3.10.0"
ipykernel = "^6.29.5"
numpy = "^2.2.1"
streamlit = "^1.41.1"
opencv-python = "^4.10.0.84"
scikit-learn = "^1.6.0"
pillow = "^11.0.0"
pandas = "^2.2.3"
diffusers = "^0.32.2"
torch = "^2.6.0"
torchvision = "^0.21.0"
datasets = "^3.3.2"
accelerate = "^1.4.0"
torchmetrics = {extras = ["image"], version = "^1.7.0"}
huggingface-hub = {extras = ["cli"], version = "^0.29.2"}
tensorboard = "^2.19.0"
wandb = "^0.19.8"
dotenv = "^0.9.9"


[tool.poetry.group.dev.dependencies]
notebook = "^7.3.2"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
